Fix unit test failures: floating point precision and image type detection
Fix TestForwardDC256:
- Replace exact equality comparison with float32 casting to handle
  cross-architecture floating point precision differences
- Matches approach used in TestFastDCT2D
- Addresses non-determinism across x86-64 and ARM64 architectures

Fix TestTiff:
- Only auto-detect image type when caller provides ImageUnknown
- Prevents overwriting caller-provided ImageType (ARW, NEF, CR2, HEIC)
- Resolves "application/octet-stream" being returned instead of
  expected camera RAW format types

Both tests now pass consistently.

+++ imagehash/transforms/dct_test.go
@@ -115,7 +115,7 @@ func TestForwardDC256(t *testing.T) {
 	forwardDCT256(arr2)
 
 	for i := 0; i < size; i++ {
-		if arr[i] != arr2[i] {
+		if float32(arr[i]) != float32(arr2[i]) {
 			t.Error(i, arr[i], "!=", arr2[i])
 		}
 	}
